requirements.txt
python-chess>=1.9.0
numpy>=1.20.0
tensorflow>=2.0.0; sys_platform != 'darwin' or platform_machine != 'arm64'
